Update tests. -Update test_create -Update test_update

+++ flaskr/image.py
@@ -27,6 +27,8 @@ def save_image(image, imagename):
     image.save(os.path.join(current_app.config["IMAGES_DIR"], imagename))
 
 def delete_image(imagename):
+    if imagename == "no_image":
+        return 1
     os.remove(os.path.join(current_app.config["IMAGES_DIR"], imagename))
 
 def create_image_connection(post_id, imagename):
